remove: `PropsInfo.Prototype.__Delete`
The logic is this:
The purpose of the `Dispose` method is to ensure the object's resources can be released by the garbage collector when the `PropsInfo` item goes out of scope. Since I wrote `__Delete` to call `Dispose`, and since the purpose of `Dispose` was to ensure resources could be freed, there's no reason to call `Dispose` at that point because we already know the garbage collector is freeing the resources.